backend/requirements.txt
aiofiles==0.8.0
anyio==3.6.1
click==8.1.3
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
psycopg2-binary==2.9.3
pydantic==1.9.1
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
